Задание 5 . Передел фактически все и ордер и заказ и таблиц

+++ src/main/java/com/example/springshop/dto/ProductDto.java
@@ -11,7 +11,7 @@ public class ProductDto {
     private Long id;
     private String name;
     private Integer price;
-    private Double count;
+    private Integer count;
 
     public void incrementCount() {
         this.count++;
